requirements.txt
streamlit
ffmpeg-python
click
matplotlib
numpy
scipy
scikit-image
torch
torchvision
tqdm
wget
Cython
opencv-python
pydicom
certifi
cycler
decorator
imageio
joblib
kiwisolver
matplotlib
networkx
numpy
opencv-python
pandas
Pillow
pyparsing
python-dateutil
pytz
PyWavelets
scikit-image
scikit-learn
scipy
six
threadpoolctl
tifffile
torch
torchvision
tqdm
TypeGuard
typing-extensions
pandas